Nuke the requirement to use Kube.
Rip the hard requirement for Kube out of the
project. This is INCREDIBLY messy at the moment
but I'm putting this somewhere remote in case
my PC explodes.

Some notes:
1. Building containers atm is a mess, and you have to
follow a bit of the development guide to build *some*
of the images.
2. The images that are missing are then an evaluator
and a matchfunction. The default evaluator inside
the cmd/default-evaluator process can be built
(call it "open-match-evaluator) and I hacked one
of the tutorials to get a matchfunction working -
tutorials/default_evaluator/matchfunction (build
it as "open-match-matchfunction")

There's a bunch of insanity in addition.